Updated generator Echo and Empty bots
