--- Generator/generator-botbuilder-java/generators/app/templates/echo/project/deploymentTemplates/new-rg-parameters.json
@@ -1,43 +1,0 @@
-{
-  "$schema":
-    "https://schema.management.azure.com/schemas/2015-01-01/deploymentParameters.json#",
-  "contentVersion": "1.0.0.0",
-  "parameters": {
-    "groupLocation": {
-      "value": ""
-    },
-    "groupName": {
-      "value": ""
-    },
-    "appId": {
-      "value": ""
-    },
-    "appSecret": {
-      "value": ""
-    },
-    "botId": {
-      "value": ""
-    },
-    "botSku": {
-      "value": ""
-    },
-    "newAppServicePlanName": {
-      "value": ""
-    },
-    "newAppServicePlanSku": {
-      "value": {
-        "name": "S1",
-        "tier": "Standard",
-        "size": "S1",
-        "family": "S",
-        "capacity": 1
-      }
-    },
-    "newAppServicePlanLocation": {
-      "value": ""
-    },
-    "newWebAppName": {
-      "value": ""
-    }
-  }
-}
