+++ Generator/generator-botbuilder-java/generators/app/templates/echo/src/main/java/EchoBot.java
@@ -9,7 +9,6 @@ import com.microsoft.bot.builder.MessageFactory;
 import com.microsoft.bot.builder.TurnContext;
 import com.microsoft.bot.schema.ChannelAccount;
 import org.apache.commons.lang3.StringUtils;
-import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -23,7 +22,6 @@ import java.util.concurrent.CompletableFuture;
  * #onMembersAdded(List, TurnContext)} will send a greeting to new conversation participants.
  * </p>
  */
-@Component
 public class EchoBot extends ActivityHandler {
 
     @Override
