--- Generator/generator-botbuilder-java/generators/app/templates/echo/src/test/java/ApplicationTest.java
@@ -1,19 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package <%= packageName %>;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ApplicationTest {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
