+++ Generator/generator-botbuilder-java/generators/app/templates/empty/project/README.md
@@ -32,7 +32,7 @@ This sample is a Spring Boot app and uses the Azure CLI and azure-webapp Maven p
 As described on [Deploy your bot](https://docs.microsoft.com/en-us/azure/bot-service/bot-builder-deploy-az-cli), you will perform the first 4 steps to setup the Azure app, then deploy the code using the azure-webapp Maven plugin.
 
 ### 1. Login to Azure
-From a command (or PowerShell) prompt in the root of the bot folder, execute:
+From a command (or PowerShell) prompt in the root of the bot folder, execute:  
 `az login`
 
 ### 2. Set the subscription
@@ -45,7 +45,7 @@ If you aren't sure which subscription to use for deploying the bot, you can view
 
 Replace `<botname>` and `<appsecret>` with your own values.
 
-`<botname>` is the unique name of your bot.
+`<botname>` is the unique name of your bot.  
 `<appsecret>` is a minimum 16 character password for your bot.
 
 Record the `appid` from the returned JSON
@@ -54,15 +54,15 @@ Record the `appid` from the returned JSON
 Replace the values for `<appid>`, `<appsecret>`, `<botname>`, and `<groupname>` in the following commands:
 
 #### To a new Resource Group
-`az deployment create --name "echoBotDeploy" --location "westus" --template-file ".\deploymentTemplates\template-with-new-rg.json" --parameters groupName="<groupname>" botId="<botname>" appId="<appid>" appSecret="<appsecret>"`
+`az deployment sub create --name "echoBotDeploy" --location "westus" --template-file ".\deploymentTemplates\template-with-new-rg.json" --parameters appId="<appid>" appSecret="<appsecret>" botId="<botname>" botSku=S1 newAppServicePlanName="echoBotPlan" newWebAppName="echoBot" groupLocation="westus" newAppServicePlanLocation="westus"`
 
 #### To an existing Resource Group
-`az group deployment create --name "echoBotDeploy" --resource-group "<groupname>" --template-file ".\deploymentTemplates\template-with-preexisting-rg.json" --parameters botId="<botname>" appId="<appid>" appSecret="<appsecret>"`
+`az deployment group create --resource-group "<groupname>" --template-file ".\deploymentTemplates\template-with-preexisting-rg.json" --parameters appId="<appid>" appSecret="<appsecret>" botId="<botname>" newWebAppName="echoBot" newAppServicePlanName="echoBotPlan" appServicePlanLocation="westus" --name "echoBot"`
 
 ### 5. Update app id and password
 In src/main/resources/application.properties update
-  - `MicrosoftAppPassword` with the botsecret value
-  - `MicrosoftAppId` with the appid from the first step
+- `MicrosoftAppPassword` with the botsecret value
+- `MicrosoftAppId` with the appid from the first step
 
 ### 6. Deploy the code
 - Execute `mvn clean package`
